Move code from retroarch.c to command.c
